Added Logos
Added section that serves as a Table of Contents - includes HTML -
Structure - CSS - Python links to different portions of the page. My
plans are to make a fixed nav bar of the ToC when the user scrolls past
it on a portion of the page. Idea is to have the ToC available within
“reach” instead of scrolling back to the top to navigate to the next
section.